console karaf: remove maven, xtext, groovy nature
Xtext and Groovy is not used by this bundle at all.
The Maven nature is not necessary at all. If you want to do some
development for this bundle it is simpler to add the respective Karaf
dependency bundle to your Eclipse IDE. The Maven Nature triggers the
Eclipse IDE to require some lifecycle plugins, that are not needed at
all.
As long as Eclipse SmartHome uses Eclipse PlugIns and Tycho, there is no
need to add all the m2e support plugins to the Eclipse IDE.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>